requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
openai==1.3.7
sentence-transformers==2.7.0
faiss-cpu==1.7.4
markdown==3.5.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
streamlit==1.28.1
requests==2.31.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
scikit-learn==1.3.2
plotly==5.17.0
openpyxl==3.1.2
